qwc-postgrest-auth/requirements.txt
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask_login==0.5.0
Flask-WTF==1.0.0
python-i18n==0.3.9
Werkzeug==2.0.2
requests==2.26.0
#qwc-services-core==1.3.7
geon-services-core
